+++ src/store/AuthStore.tsx
@@ -2,6 +2,7 @@ import { create } from "zustand";
 import { persist } from "zustand/middleware";
 import { postSignIn } from "@/api/auth";
 import { postProfile } from "@/api/profile";
+import { AxiosError } from "axios";
 
 interface Profile {
   updatedAt: string;
@@ -38,7 +39,7 @@ interface AuthStore {
   accessToken: string | null;
   refreshToken: string | null;
   isLoggedIn: boolean;
-  login: (email: string, password: string) => Promise<void>;
+  login: (email: string, password: string) => Promise<string | undefined>; //반환 타입 수정
   logout: () => void;
   setAccessToken: (token: string) => void; // Access Token을 업데이트하는 함수 추가
   updateProfile: (
@@ -55,16 +56,22 @@ const useAuthStore = create(
       refreshToken: null,
       isLoggedIn: false,
       login: async (email, password) => {
-        const userData = await postSignIn({ email, password });
-        if (userData) {
-          set({
-            user: userData.user,
-            accessToken: userData.accessToken,
-            refreshToken: userData.refreshToken,
-            isLoggedIn: true,
-          });
-          localStorage.setItem("accessToken", userData.accessToken);
-          localStorage.setItem("refreshToken", userData.refreshToken);
+        try {
+          const userData = await postSignIn({ email, password });
+          if (userData) {
+            set({
+              user: userData.user,
+              accessToken: userData.accessToken,
+              refreshToken: userData.refreshToken,
+              isLoggedIn: true,
+            });
+            localStorage.setItem("accessToken", userData.accessToken);
+            localStorage.setItem("refreshToken", userData.refreshToken);
+          }
+        } catch (error) {
+          if (error instanceof AxiosError) {
+            return error.response?.data.message;
+          }
         }
       },
       logout: () => {
